basic accounts support
This is pretty primitive...
I removed the ability to change the statements dir.
For now, the statements must be in the same dir as the account.
This greatly simplifies some of the logic.

Add support for a list of "null_statements" and a list of
"known_missing_statements".
(The hope is that I can alert on the later at some point.)

Update requirements.