override statements directory, "filename_date_map"
Support overriding the statements directory name,
Also, remove the hacky setting of "pay-stubs" as the dir name
if the type was "payroll.
(That was pretty stupid...)

Add support for "filename_date_map".
This allows using the "raw" filename as defined by the vendor
themselves.
Rather than trying to parse PDFs, just add manual way to say a given
filename maps to a given day.